# docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "TaskMate API",
    "description": "This API provides a comprehensive set of features for managing your tasks. You can create, view, edit, and delete tasks, as well as filter and sort them based on various criteria. The API also supports marking tasks as completed and setting their priority.",
    "version": "1.0.0",
    "contact": {
      "email": "n11422807@qut.edu.au"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths": {
    "/users/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "description": "User details for registration",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request or user already exists"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "User login",
        "requestBody": {
          "description": "User login credentials",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/users/tasks": {
      "get": {
        "summary": "Get all tasks",
        "security": [{
          "bearerAuth": []
        }],
        "responses": {
          "200": {
            "description": "A list of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "security": [{
          "bearerAuth": []
        }],
        "requestBody": {
          "description": "Task to be created",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/users/tasks/todo": {
      "get": {
        "summary": "Get all todo tasks",
        "security": [{
          "bearerAuth": []
        }],
        "responses": {
          "200": {
            "description": "A list of todo tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/users/tasks/completed": {
      "get": {
        "summary": "Get all completed tasks",
        "security": [{
          "bearerAuth": []
        }],
        "responses": {
          "200": {
            "description": "A list of completed tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/users/tasks/date/{date}": {
      "get": {
        "summary": "Get tasks by date",
        "security": [{
          "bearerAuth": []
        }],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "description": "The date of the tasks to retrieve",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of tasks for the specified date",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/users/tasks/{id}": {
      "get": {
        "summary": "Get a specific task by id",
        "security": [{
          "bearerAuth": []
        }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the task to retrieve",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Task not found"
          }
        }
      },
      "put": {
        "summary": "Update a task",
        "security": [{
          "bearerAuth": []
        }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the task to update",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Task details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Task not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "security": [{
          "bearerAuth": []
        }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the task to delete",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/users/tasks/{id}/completed": {
      "put": {
        "summary": "Update a task's completed status",
        "security": [{
          "bearerAuth": []
        }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the task to update",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Task completion status",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "completed": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task completed status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserRegister": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["email", "username", "password"]
      },
      "UserRegisterResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "msg": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "UserLogin": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["email", "password"]
      },
      "UserLoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "completed": {
            "type": "boolean"
          },
          "date": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "TaskCreation": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "completed": {
            "type": "boolean"
          },
          "date": {
            "type": "string",
            "format": "date"
          }
        },
        "required": ["title", "color", "priority", "date"]
      }
    }
  }
}